2004-06-18  Jay Rossiter <cryptographite@users.sf.net>  1.4.4-cvs-b199
        * Bug 969046: Prevent an impatient user from deleting two albums
          at once by clicking the form submit twice.


git-svn-id: file:///home/jmullan/src/gallery-git/repo/trunk/gallery@6386 57fcd75e-5312-0410-8df3-f5eb6fbb1595

+++ delete_album.php
@@ -34,7 +34,9 @@ doctype();
 echo "\n<html>";
 
 if (!empty($delete)) {
-	$gallery->album->delete();
+	if ($guid == $gallery->album->fields['guid']) {
+		$gallery->album->delete();
+	}
 	dismissAndReload();
 	return;
 }
@@ -56,6 +58,7 @@ if ($gallery->album) {
 <b><?php echo $gallery->album->fields["title"] ?></b>
 <p>
 <?php echo makeFormIntro("delete_album.php"); ?>
+<input type="hidden" name="guid" value="<?php echo $gallery->album->fields['guid']; ?>">
 <input type="submit" name="delete" value="<?php echo _("Delete") ?>">
 <input type="button" name="cancel" value="<?php echo _("Cancel") ?>" onclick='parent.close()'>
 </form>
